refactor to simplify main.cpp
